fix: convert landing page to server components — extract HeroSection client island
Remove \"use client\" from page.tsx. The hero section (with framer-motion
animations) is extracted into a HeroSection client component. All other
sections already use client components (RevealSection, TerminalComponents,
BcHomeDemo, etc.) that have their own \"use client\" directives.

Built HTML now contains static content (feature text, terminal mocks)
without requiring JavaScript to render.

Closes #2124
